chore(stella-pipeline): split replay/verify test modules under the god-file limit; wire schema + baseline
