+++ crates/stella-cli/src/settings/toml_config.rs
@@ -193,6 +193,8 @@ pub struct AgentsSection {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub pipeline_require_diff_coverage: Option<Toggle>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub pipeline_require_independent_verifier: Option<Toggle>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     pub model_timeout_secs: Option<u64>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub compaction_budget_tokens: Option<u64>,
@@ -597,6 +599,7 @@ pub fn raise_agents(cfg: &AgentEngineConfig) -> (AgentsSection, ModelsSection) {
         pipeline_candidates: cfg.pipeline_candidates,
         pipeline_verifier_evidence_demand: cfg.pipeline_verifier_evidence_demand,
         pipeline_require_diff_coverage: cfg.pipeline_require_diff_coverage,
+        pipeline_require_independent_verifier: cfg.pipeline_require_independent_verifier,
         model_timeout_secs: cfg.model_timeout_secs,
         compaction_budget_tokens: cfg.compaction_budget_tokens,
         tool_result_horizon_steps: cfg.tool_result_horizon_steps,
@@ -670,6 +673,7 @@ fn lower_agents(agents: AgentsSection, models: ModelsSection) -> Option<AgentEng
         pipeline_candidates: agents.pipeline_candidates,
         pipeline_verifier_evidence_demand: agents.pipeline_verifier_evidence_demand,
         pipeline_require_diff_coverage: agents.pipeline_require_diff_coverage,
+        pipeline_require_independent_verifier: agents.pipeline_require_independent_verifier,
         model_timeout_secs: agents.model_timeout_secs,
         compaction_budget_tokens: agents.compaction_budget_tokens,
         tool_result_horizon_steps: agents.tool_result_horizon_steps,
